requirements.txt
numpy
Pillow
matplotlib
tqdm
torch
tensorboard
